+++ release/oci-standard/selftest.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+
+# Fail on error.
+set -e
+
+# Display all commands.
+# set -x
+
+echo "Invoking CrateDB MCP Server"
+cratedb-mcp --version