+++ CHANGES.md
@@ -5,6 +5,7 @@
 - Parameters: Added CLI option `--path` and environment variable `CRATEDB_MCP_PATH`
 - Dependencies: Allowed updating to [FastMCP 2.10], which includes a specification
   update to [MCP spec 2025-06-18].
+- OCI: Added building OCI standard image `cratedb-mcp`
 
 [FastMCP 2.10]: https://github.com/jlowin/fastmcp/releases/tag/v2.10.0
 [MCP spec 2025-06-18]: https://modelcontextprotocol.io/specification/2025-06-18/changelog